+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/identity/IdentityTaskHelper.kt
@@ -18,14 +18,13 @@ package org.matrix.android.sdk.internal.session.identity
 
 import org.matrix.android.sdk.api.session.identity.IdentityServiceError
 import org.matrix.android.sdk.internal.network.executeRequest
-import org.matrix.android.sdk.internal.session.identity.model.IdentityAccountResponse
 
 internal suspend fun getIdentityApiAndEnsureTerms(identityApiProvider: IdentityApiProvider, userId: String): IdentityAPI {
     val identityAPI = identityApiProvider.identityApi ?: throw IdentityServiceError.NoIdentityServerConfigured
 
     // Always check that we have access to the service (regarding terms)
-    val identityAccountResponse = executeRequest<IdentityAccountResponse>(null) {
-        apiCall = identityAPI.getAccount()
+    val identityAccountResponse = executeRequest(null) {
+        identityAPI.getAccount()
     }
 
     assert(userId == identityAccountResponse.userId)