+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/timeline/PaginationTask.kt
@@ -42,10 +42,10 @@ internal class DefaultPaginationTask @Inject constructor(
 
     override suspend fun execute(params: PaginationTask.Params): TokenChunkEventPersistor.Result {
         val filter = filterRepository.getRoomFilter()
-        val chunk = executeRequest<PaginationResponse>(globalErrorReceiver) {
-            isRetryable = true
-            apiCall = roomAPI.getRoomMessagesFrom(params.roomId, params.from, params.direction.value, params.limit, filter)
-        }
+        val chunk = executeRequest(globalErrorReceiver,
+                { roomAPI.getRoomMessagesFrom(params.roomId, params.from, params.direction.value, params.limit, filter) },
+                { isRetryable = true }
+        )
         return tokenChunkEventPersistor.insertInDb(chunk, params.roomId, params.direction)
     }
 }